# openapi.yaml
openapi: 3.0.3
info:
  title: Beacon API
  version: "1.0.0"
  description: |-
    Beacon is a crisis-damage crowdsourcing backend: anonymous mobile reporters
    submit building-damage reports (identified only by an `X-Device-Id` header),
    and authenticated analysts (JWT bearer) verify, dispatch and export them.

    ## Visibility tiers
    - **Anonymous / public** callers (no JWT) and the authenticated read-only
      `external_viewer` role see **verified reports only**, in a locked-down
      *public projection*: coordinates coarsened to 3 decimal places (~110 m),
      and `submitterId`, `what3words`, `plusCode`, `landmark`, `buildingId`,
      `buildingSource`, `infraName`, `gpsAccuracyMeters`, `description`,
      tasking/operational fields, `modular` and `anonymization` are stripped.
      Unverified reports return 404 so their existence and position are never
      leaked. Public aggregates (`/reports/area-groups`) count verified
      reports only for this tier, coherent with the verified-only public map.
    - **Analyst roles** (`field_validator`, `co_analyst`, `regional_analyst`,
      `crisis_admin`) see all statuses at full precision, limited to the
      crises in their `crisisScope` (`"*"` = org-wide).

    ## Crisis scoping
    Where `crisisId` is optional (stats, export), an omitted value resolves to
    the **newest active crisis** (the same crisis `GET /v1/crises/active`
    returns), falling back to a seeded default only when no crisis is active.

...

...

...
servers:
...
paths:
...
  /v1/reports/export:
    get:
      tags: [exports]
      operationId: exportReports
      summary: Bulk export (GeoJSON / HXL CSV / GeoPackage / KML / PDNA pivot)
      description: |-
        Analyst-only bulk export with exact coordinates and the full
        per-report row. The read-only `external_viewer` role is **denied
        (403)** — its read need is served by the coarsened public map
        endpoints instead. `crisisId` defaults to the newest active crisis.

        **Row formats (geojson / csv / gpkg / kml)** carry the gate field
        list: `damage_classification` ∈ {Minimal, Partial, Complete} (the
        canonical 3-tier rollup), raw `damage` grade, `infrastructure_type`,
        `infrastructure_name`, `hazard_type` (empty when the reporter asserted
        none), `timestamp` (ISO-8601 of capturedAt, UTC), and the modular
        secondary-impact sections flattened to snake_case — the three known
        sections under their stable names (`electricity`, `health_services`,
        `pressing_needs`, always present) plus any later-added section
        DYNAMICALLY (e.g. `pressingNeedsOther` → `pressing_needs_other`,
        appended after the fixed CSV columns) — plus `possiblyDamaged`,
        `debris`, `buildingId`, `verification`, `place`, `description`
        (free-text note, analyst language), `plus_code`, `accuracy_m`, and
        the `admin1_shapeid` / `admin2_shapeid` / `admin3_shapeid` columns
        (geoBoundaries shapeIDs / seed codes — NOT OCHA P-codes; their HXL
        tags use `+id`, never `+code`).

        - `geojson` — FeatureCollection; geometry is a Point `[lng, lat]` in
          decimal degrees, or `null` for a location-unresolved report (never
          `[0,0]`).
        - `csv` — header row + an HXL hashtag row, `latitude`/`longitude`
          columns blank for unresolved reports.
        - `gpkg` — OGC GeoPackage (single SQLite file), NULL geometry for
          unresolved reports.
        - `kml` — one Placemark per *resolved* report (unresolved skipped).
        - `pdna` — a different shape: a PDNA-ready **damage-count** pivot
          (sector × admin-2) of report counts per grade, HXL-tagged CSV. The
          3-tier rollup (minimal/partial/complete) sums to `total`; the
          5-level EMS-98 columns are detail only. It is NOT a loss/cost
          estimation (no monetary figures).
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [geojson, csv, gpkg, kml, pdna]
            default: geojson
        - name: crisisId
          in: query
          schema: { type: string }
          description: Defaults to the newest active crisis when omitted.
        - name: damage
          in: query
          schema: { type: string }
          description: Comma-separated or repeated damage grades.
        - name: verification
          in: query
          schema: { type: string }
          description: "Comma-separated or repeated: pending,verified,flagged."
        - name: q
          in: query
          schema: { type: string }
        - name: adm2Pcode
          in: query
          schema: { type: string }
        - name: adm3Pcode
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/BBox'
      responses:
        "200":
          description: The export file (Content-Disposition attachment).
          content:
            application/geo+json:
              schema:
                type: object
                description: GeoJSON FeatureCollection (format=geojson).
            text/csv:
              schema:
                type: string
                description: HXL-tagged CSV (format=csv) or PDNA damage-count pivot (format=pdna).
            application/geopackage+sqlite3:
              schema:
                type: string
                format: binary
                description: OGC GeoPackage (format=gpkg).
            application/vnd.google-earth.kml+xml:
              schema:
                type: string
                description: KML document (format=kml).
        "400":
          description: Unknown format (`bad_format`) or bad bbox (`bad_bbox`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: external_viewer cannot export raw data (`forbidden`), or the crisis is out of the caller's scope (`out_of_scope`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/crises:
    get:
      tags: [crises]
      operationId: listCrises
      summary: List crises (optionally by status)
      parameters:
        - name: status
          in: query
          schema: { type: string }
          description: "Comma-separated lifecycle filter, e.g. `active,proposed`. Statuses: active, proposed, closed, dismissed."
      responses:
        "200":
          description: Crises.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Crisis'
  /v1/crises/active:
    get:
      tags: [crises]
      operationId: activeCrisis
      summary: The newest active crisis
      description: Returns the newest active crisis, or 204 when no crisis is active.
      responses:
        "200":
          description: The active crisis.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crisis'
        "204":
          description: No active crisis.
  /v1/crises/near:
    get:
      tags: [crises]
      operationId: nearbyCrises
      summary: Active/proposed crises near a point (location-first launch)
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, minimum: -90, maximum: 90 }
        - name: lng
          in: query
          required: true
          schema: { type: number, minimum: -180, maximum: 180 }
        - name: radiusKm
          in: query
          schema: { type: number, default: 50 }
      responses:
        "200":
          description: Nearby crises with distance (and whether the point is covered).
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Crisis'
                  count:
                    type: integer
        "400":
          $ref: '#/components/responses/BadRequest'
  /v1/crises/{id}:
    get:
      tags: [crises]
      operationId: getCrisis
      summary: One crisis by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The crisis.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crisis'
        "404":
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Analyst JWT issued by POST /v1/auth/login.
  parameters:
...
    BBox:
      name: bbox
      in: query
      required: false
      schema: { type: string }
      description: "Geographic filter: `minLng,minLat,maxLng,maxLat` (WGS84 decimal degrees)."
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Analyst login required (missing/invalid bearer token).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found (also returned for resources the caller's tier may not see).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Stable machine-readable code (e.g. validation, bad_json, unauthorized, out_of_scope, not_found, rate_limited).
        message:
          type: string
...
    Crisis:
      type: object
      description: A crisis event. Emits both lat/lng (dashboard) and centerLat/centerLng (mobile).
      properties:
        id: { type: string }
        title: { type: string }
        area: { type: string }
        nature:
          type: string
          description: e.g. earthquake, flood, tsunami, hurricane, wildfire, explosion, chemical, conflict, civil_unrest.
        lat: { type: number }
        lng: { type: number }
        centerLat: { type: number }
        centerLng: { type: number }
        source: { type: string }
        startedAt:
          type: string
          format: date-time
        startedAgoHrs: { type: integer }
        glide:
          type: string
          nullable: true
          description: Cross-org disaster event key (GLIDE).
        responseLevel:
          type: integer
          nullable: true
          description: UNDP corporate crisis Level 1/2/3.
        radiusKm:
          type: number
          description: Coverage radius around the center point (km).
        endedAt:
          type: string
          format: date-time
          nullable: true
          description: Null while ongoing.
        status:
          type: string
          enum: [active, proposed, closed, dismissed]
          description: proposed = emergent cluster awaiting analyst confirmation.
        responseId:
          type: string
          nullable: true
        reportCount: { type: integer }
        distanceKm:
          type: number
          nullable: true
          description: Only on /crises/near responses (km from the queried point).
        covers:
          type: boolean
          nullable: true
          description: Only on /crises/near — true if the queried point is within radiusKm.
...